feat: add ideia about stoke

+++ Clininca_veterinaria_2023_DDD/WebApi/Model/ProdutoEstoqueDto.cs
@@ -2,7 +2,7 @@
 {
     public class ProdutoEstoqueDto
     {
-        public long IdProduto { get; set; }
+        public int IdProduto { get; set; }
         public int ID_Usuario { get; set; }
         public long IdEstoque { get; set; }
         public DateTime? DataEntrada { get; set; }
